+++ src/lib.rs
@@ -1,6 +1,5 @@
 //! Concise trait implementations.
 
-#![no_std]
 #![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, nonstandard_style)]
 #![warn(future_incompatible)]