+++ src/macros.rs
@@ -91,9 +91,9 @@ macro_rules! impl_as_mut {
 /// [`Deref`]: std::ops::Deref
 #[macro_export]
 macro_rules! impl_deref {
-    ($ty:ty, $inner:ty) => {
+    ($ty:ty, $target:ty) => {
         impl ::core::ops::Deref for $ty {
-            type Target = $inner;
+            type Target = $target;
 
             fn deref(&self) -> &Self::Target {
                 &self.0
@@ -101,9 +101,9 @@ macro_rules! impl_deref {
         }
     };
 
-    ($ty:ty, $inner:ty, $field:ident) => {
+    ($ty:ty, $target:ty, $field:ident) => {
         impl ::core::ops::Deref for $ty {
-            type Target = $inner;
+            type Target = $target;
 
             fn deref(&self) -> &Self::Target {
                 ::core::ops::Deref::deref(&self.$field)
@@ -145,6 +145,41 @@ macro_rules! impl_deref_mut {
     };
 
     ($ty:ty, $field:ident) => {
+        impl ::core::ops::DerefMut for $ty {
+            fn deref_mut(&mut self) -> &mut Self::Target {
+                &mut self.$field
+            }
+        }
+    };
+}
+
+#[macro_export]
+macro_rules! impl_deref_and_mut {
+    ($ty:ty, $target:ty) => {
+        impl ::core::ops::Deref for $ty {
+            type Target = $target;
+
+            fn deref(&self) -> &Self::Target {
+                &self.0
+            }
+        }
+
+        impl ::core::ops::DerefMut for $ty {
+            fn deref_mut(&mut self) -> &mut Self::Target {
+                &mut self.0
+            }
+        }
+    };
+
+    ($ty:ty, $target:ty, $field:ident) => {
+        impl ::core::ops::Deref for $ty {
+            type Target = $target;
+
+            fn deref(&self) -> &Self::Target {
+                ::core::ops::Deref::deref(&self.$field)
+            }
+        }
+
         impl ::core::ops::DerefMut for $ty {
             fn deref_mut(&mut self) -> &mut Self::Target {
                 &mut self.$field
@@ -203,7 +238,7 @@ macro_rules! impl_into {
     };
 }
 
-/// Implement [`Display`] for enums using a static string for each variant.
+/// Implement [`Display`] for enums using a static string or format args for each variant.
 ///
 /// # Examples
 /// ```
@@ -218,6 +253,16 @@ macro_rules! impl_into {
 ///
 /// assert_eq!(Foo::Bar.to_string(), "bar");
 /// assert_eq!(Foo::Qux.to_string(), "qux");
+///
+/// enum CoordOrMsg {
+///     Coord(i64, i64),
+///     Msg(&'static str),
+/// }
+///
+/// impl_display_enum!(CoordOrMsg, Coord(x, y) => "{x}, {y}", Msg(msg) => "message: {msg}");
+///
+/// assert_eq!(CoordOrMsg::Coord(4, 2).to_string(), "4, 2");
+/// assert_eq!(CoordOrMsg::Msg("hi").to_string(), "message: hi");
 /// ```
 ///
 /// [`Display`]: std::fmt::Display
@@ -235,9 +280,61 @@ macro_rules! impl_display_enum {
         }
     };
 
-    ($ty:ty, $($variant:ident => $stringified:literal),+,) => {
+    ($ty:ty, $($variant:ident => $stringified:literal),+ ,) => {
         impl_display_enum!($ty, $($variant => $stringified),+)
-    }
+    };
+
+    ($ty:ty, $($variant:ident ($($inner:ident),+) => $format:literal),+) => {
+        impl ::core::fmt::Display for $ty {
+            fn fmt(&self, fmt: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
+                use ::core::fmt::Write as _;
+
+                // a more efficient method (format_args) is blocked by:
+                // https://github.com/rust-lang/rust/issues/15023
+                let mut buf = ::std::string::String::new();
+
+                match self {
+                    $(
+                        Self::$variant($($inner),+) =>
+                            ::core::write!(&mut buf, $format, $($inner = $inner),+)?,
+                    )*
+                };
+
+                fmt.write_str(&buf)
+            }
+        }
+    };
+
+    ($ty:ty, $($variant:ident ($($inner:ident),+) => $format:literal),+ ,) => {
+        impl_display_enum!($ty, $($variant ($($inner),+) => $format),+)
+    };
+
+    ($ty:ty, $($variant:ident { $($inner:ident),+ } => $format:literal),+) => {
+        impl ::core::fmt::Display for $ty {
+            fn fmt(&self, fmt: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
+                use ::core::fmt::Write as _;
+
+                // a more efficient method (format_args) is blocked by:
+                // https://github.com/rust-lang/rust/issues/15023
+                let mut buf = ::std::string::String::new();
+
+                match self {
+                    $(
+                        Self::$variant { $($inner),+ } =>
+                            ::core::write!(&mut buf, $format, $($inner = $inner),+)?,
+                    )*
+                };
+
+                fmt.write_str(&buf)
+            }
+        }
+    };
+
+    // TODO: mixed named and positional variant support
+
+    ($ty:ty, $($variant:ident { $($inner:ident),+ } => $format:literal),+ ,) => {
+        impl_display_enum!($ty, $($variant ($($inner),+) => $format),+)
+    };
 }
 
 #[cfg(test)]
@@ -260,7 +357,9 @@ mod tests {
     impl_into!(Foo, String);
 
     static_assertions::assert_impl_all!(
-        Foo: AsRef<String>,
+        Foo:
+        // impls
+        AsRef<String>,
         AsMut<String>,
         Deref<Target = String>,
         DerefMut,
@@ -279,29 +378,50 @@ mod tests {
     impl_deref!(Bar, String, foo);
     impl_deref_mut!(Bar, foo);
 
-    // impl_from!(Bar, String);
-    // impl_into!(Bar, String);
+    static_assertions::assert_impl_all!(
+        Bar:
+        // impls
+        AsRef<Foo>,
+        AsMut<Foo>,
+        Deref<Target = String>,
+        DerefMut,
+    );
 
+    #[allow(dead_code)]
     #[test]
     fn impl_display() {
-        #[derive(Debug)]
         enum Foo {
             Bar,
             Qux,
         }
-
         impl_display_enum!(Foo, Bar => "bar", Qux => "qux");
-
         assert_eq!(Foo::Bar.to_string(), "bar");
         assert_eq!(Foo::Qux.to_string(), "qux");
 
-        #[allow(dead_code)]
-        #[derive(Debug)]
         enum FooComma {
             Bar,
             Qux,
         }
-
         impl_display_enum!(FooComma, Bar => "bar", Qux => "qux",);
+
+        enum FooContents {
+            Bar(u64, u64),
+        }
+        impl_display_enum!(FooContents, Bar (x, y) => "x: {x}; y: {y}");
+        assert_eq!(FooContents::Bar(4, 2).to_string(), "x: 4; y: 2");
+
+        enum FooContents2 {
+            Qux { msg: &'static str },
+        }
+        impl_display_enum!(FooContents2, Qux { msg } => "{msg}");
+        assert_eq!(FooContents2::Qux { msg: "foo" }.to_string(), "x: 4; y: 2");
+
+        // enum FooContents3 {
+        //     Bar(u64, u64),
+        //     Qux { msg: &'static str },
+        // }
+        // impl_display_enum!(FooContents3, Bar (x, y) => "x: {x}; y: {y}", Qux { msg } => "{msg}");
+        // assert_eq!(FooContents3::Bar(4, 2).to_string(), "x: 4; y: 2");
+        // assert_eq!(FooContents3::Qux { msg: "foo" }.to_string(), "x: 4; y: 2");
     }
 }