+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `AsRef`, `AsMut`, `Deref`, and `DerefMut` invocations for structs other than newtypes.
+- Add support in `impl_display_enum!` for variables interpolation.
 
 ## 0.0.2 - 2022-09-25