+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add combined `impl_as_deref_and_mut!` macro.
+- Add `forward_deref_and_mut!` macro.
 - Add `impl_as_ref!`, `impl_as_mut!`, `impl_deref!`, and `impl_deref_mut!` invocations for structs other than newtypes.
 - Add support in `impl_display_enum!` for variables interpolation.